Refactored headers, added inlines, moved all gtests under common entry point and created separate CMake function for automatic tests resolution
