+++ cmake/build-utils.cmake
@@ -10,7 +10,7 @@ function(add_all_subdirs)
     endforeach()
 endfunction()
 
-function(list_all_tests name_pattern)
+function(list_all_tests_separate name_pattern)
     file (GLOB_RECURSE TEST_SOURCES ${name_pattern})
 
     foreach (TEST_SRC ${TEST_SOURCES})
@@ -30,4 +30,21 @@ function(list_all_tests name_pattern)
 
         message ("Loaded test sequence: ${FILENAME}")
     endforeach ()
+endfunction()
+
+function(list_all_tests name_pattern)
+    file (GLOB_RECURSE TEST_SOURCES ${name_pattern})
+    get_filename_component(FOLDER ${CMAKE_CURRENT_SOURCE_DIR} NAME_WE)
+    string (TOUPPER ${FOLDER} FOLDER)
+
+    add_executable(TEST_${FOLDER}_TARGET ${TEST_SOURCES} "entry.cpp")
+    add_test(NAME TEST_${FOLDER} COMMAND $<TARGET_FILE:TEST_${FOLDER}_TARGET>)
+    target_link_libraries(TEST_${FOLDER}_TARGET PRIVATE ${ARGN})
+
+    add_custom_command(TARGET TEST_${FOLDER}_TARGET POST_BUILD
+            COMMAND ${CMAKE_COMMAND} -E copy -t $<TARGET_FILE_DIR:TEST_${FOLDER}_TARGET> $<TARGET_RUNTIME_DLLS:TEST_${FOLDER}_TARGET>
+            COMMAND_EXPAND_LISTS
+    )
+
+    message ("Loaded test folder: TEST_${FOLDER}")
 endfunction()
