+++ libraries/events/AxonEvent.ipp
@@ -18,8 +18,7 @@ void EventSystem::AxonEventManager::subscribe(std::function<void(const T&)> call
 template<class C, typename T>
 void EventSystem::AxonEventManager::subscribe(void(C::* callback)(const T&), C* instance)
 {
-	std::type_index type = std::type_index(typeid(T));
-
+	constexpr auto type = std::type_index(typeid(T));
 	subscribers[type].push_back(
 		[callback, instance](AxonEvent* base_event) {
 			if (T* event = dynamic_cast<T*>(base_event))
