+++ libraries/networking/synapse/BasicSynapse.ipp
@@ -52,4 +52,3 @@ void Networking::BasicSynapse<NetworkController>::listen() {
 }
 
 #pragma endregion
-#pragma endregion
