--- tests/basic_checks/test_event_system.cxx
@@ -1,52 +1,0 @@
-#include <networking/AxonNetwork.hpp>
-
-#include <cassert>
-#include <cstdint>
-
-
-class TestEvent : public EventSystem::AxonEvent
-{
-private:
-	uint8_t data;
-	bool isChecked = false;
-public:
-    TestEvent() = delete;
-
-	explicit TestEvent(uint8_t data = 0)
-	{
-		this->data = data;
-	}
-
-	inline uint8_t get_data() const { return data; }
-};
-
-
-void callback(const TestEvent& event)
-{
-	uint8_t data = event.get_data();
-	assert(data == 3);
-}
-
-void callback2(const TestEvent& event)
-{
-	uint8_t data = event.get_data();
-	assert(data == 3);
-}
-
-int main()
-{
-	/* Local events */
-	TestEvent event(3);
-
-	EventSystem::AxonEventManager manager;
-
-	manager.subscribe<TestEvent>(callback);
-	manager.invoke(&event);
-
-	/* Global event manager */
-	EventSystem::GlobalEventManager& g_manager = EventSystem::GlobalEventManager::Instance();
-	g_manager.subscribe<TestEvent>(callback);
-	g_manager.subscribe<TestEvent>(callback2);
-
-	g_manager.invoke(&event);
-}
