+++ tests/gtest/test_network_pointer.cpp
@@ -1,0 +1,2 @@
+#include <gtest/gtest.h>
+
