+++ tests/gtest/test_networking.cpp
@@ -4,9 +4,3 @@ TEST(TESTS_BACKEND, TEST_BACKEND_GENERAL)
 {
 
 }
-
-int main(int argc, char* argv[])
-{
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
